backend/requirements_clean.txt
Django==5.2.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-decouple==3.8
channels==4.2.2
channels-redis==4.2.1
daphne==4.2.1
dj-database-url==2.2.0
psycopg2-binary==2.9.9
redis==5.2.1
Pillow==11.1.0
whitenoise==6.8.2
